requirements.txt
Django==4.2.27
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
clickhouse-connect==0.7.1
ruff==0.6.3
